Refactor Logout into logout.rs
