+++ epp-client/src/tests/de.rs
@@ -29,9 +29,10 @@ mod response {
     use crate::host::info::EppHostInfoResponse;
     use crate::host::update::EppHostUpdateResponse;
     use crate::login::EppLoginResponse;
+    use crate::logout::EppLogoutResponse;
     use crate::message::ack::EppMessageAckResponse;
     use crate::message::poll::EppMessagePollResponse;
-    use crate::response::{EppCommandResponseError, EppLogoutResponse};
+    use crate::response::EppCommandResponseError;
     use crate::xml::EppXml;
 
     const SVTRID: &str = "RO-6879-1627224678242975";
