+++ epp-client/src/tests/se.rs
@@ -36,9 +36,9 @@ mod request {
     use crate::host::update::HostAddRemove;
     use crate::host::update::HostChangeInfo;
     use crate::login::EppLogin;
+    use crate::logout::EppLogout;
     use crate::message::ack::EppMessageAck;
     use crate::message::poll::EppMessagePoll;
-    use crate::request::EppLogout;
     use crate::xml::EppXml;
     use chrono::{DateTime, NaiveDate};
     use std::str::FromStr;
